feat: new setting to control header transparency
Previously when welcome image was set, the header was automatically transparent with now way to disable without manual css overrides.  This reverses that, with it preserving header color even with welcome image, and adds a setting to enable transparency.

Easier to sellers to use as it's often difficult to find the correct image that won't interfere with the header menu.